requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiokafka==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.4.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
build==1.2.2.post1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
databases==0.9.0
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
deepdiff==8.0.1
Deprecated==1.2.14
distro==1.9.0
distro-info==1.0
dnspython==2.7.0
docopt==0.6.2
durationpy==0.9
elastic-transport==8.15.1
elasticsearch==8.15.1
email_validator==2.2.0
emoji==2.14.0
environs==11.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastapi-limiter==0.1.6
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
flower==2.0.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphql-core==3.2.4
groq==0.11.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
hiredis==3.0.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
humanize==4.11.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python-ng==2.2.3
keras==3.6.0
kiwisolver==1.4.7
kombu==5.4.2
kubernetes==31.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.131
libclang==18.1.1
libcst==1.4.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.51.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.13.0
orderly-set==5.2.2
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pendulum==3.0.0
pexpect==4.9.0
pika==1.3.2
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip-upgrader==1.4.15
platformdirs==4.3.6
posthog==3.7.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
pyparsing==3.1.4
pypdf==5.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
sentence-transformers==3.1.1
sentry-sdk==2.16.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
strawberry-graphql==0.246.0
streamlit==1.39.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
terminaltables==3.1.10
tf_keras==2.17.0
threadpoolctl==3.5.0
tiktoken==0.8.0
time-machine==2.16.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.15.13
unstructured-client==0.25.9
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
vine==5.1.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
